Renamed meraki_mx_uplink to meraki_mx_uplink_bandwidth
